Fix wrapping of long document bucket titles
If a document bucket had a long word in its title, for example
Pitt_Comp_Literacy_Culture_Distant_Publics_Development_Rhetoric_and_the_Subject_of_Crisis.pdf,
this word was not being wrapped and was overflowing out of the document
bucket and expanding the width of the entire page (especially on mobile).

This commit makes browsers wrap within word if necessary to prevent
overflow.

Fixes #3927