Fix search bucket content alignment
Fix the horizontal alignment of search result bucket contents on small
screens.

There was a 130px left margin inside the search result bucket contents
which, on small screens, takes up most of the screen width and squashes
the actual contents into a small space.

Make this margin disappear on small screens, and also centre the search
bucket contents within the search result bucket div that contains it,
when the search result bucket contents are not as wide as the search
result bucket div.