Rework modal focus handling for forms
The previous method of ensuring that focus remained within the field
being edited and the form's action buttons relied on asynchronously
reverting focus changes after a blur.

This had several problems:

 - There was a brief flash while focus moved to a new "disallowed"
   element and back again

 - If the page contained multiple forms, it was possible for
   the page to get into a state where focus would continually
   ping-pong back between two forms each trying to retain focus

One solution is to use FocusEvent#relatedTarget in the `blur` handler,
but this is not supported by Firefox < 48 or IE and is null in some
cases in Safari. Another would be the `focusout` event, but this
is not supported by Firefox either.

This commit instead adds a helper for implementing modal UI which makes
use of a document-level "focus" event handler.